Reorg, assem defs and some fixes
Fixed ranged animation
fixed health being coupled with 2 systems, so i separated them using events.
Re orged the scripts and some other places to be less nested and straight forward to access. prefab in animations in scripts is not good.
added asem defs (If you dont like it, deal with it)