cleaned up the beam bull shit
dont know why I duplicated it, its now located in the common beam enemy as priest inherits the beam.

The whole AI situation should be reworked to a more of a state machine or something that is more expandable without duplicating code, such as the light attack in the priest.